+++ absl/CHANGELOG.md
@@ -8,6 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com).
 
 Nothing notable unreleased.
 
+## 0.15.0 (2021-10-19)
+
+### Changed
+
+*   (testing) #128: When running bazel with its `--test_filter=` flag, it now
+    treats the filters as `unittest`'s `-k` flag in Python 3.7+.
+
 ## 0.14.1 (2021-09-30)
 
 ### Fixed